fix: resolve issue with abnormal scaling caused by zoom, and floating point line-height values

+++ src/utils/compute.ts
@@ -43,5 +43,5 @@ export function getLineHeight(element: HTMLElement) {
     // Compatible some browser
     return height * 1.1;
   }
-  return parseInt(lineHeight);
+  return parseFloat(lineHeight);
 }
